requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.7.9
click==8.2.1
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
numpy==1.26.4
pandas==2.2.2
pip==23.2.1
pluggy==1.6.0
python-multipart==0.0.20
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.0.0
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.47.1
typing_extensions==4.14.1
uvicorn==0.24.0
